Server CSV fixes #10 #9#8 #7

+++ src/main/java/g7asmt1/server/Result.java
@@ -4,12 +4,12 @@ import java.io.Serializable;
 
 public class Result implements Serializable {
     private final String serviceName;
-    private final int result;
+    private final long result;
     private final long waitingTime;
     private final long executionTime;
     private final int zone;
 
-    public Result(String serviceName, int result, long waitingTime, long executionTime, int zone) {
+    public Result(String serviceName, long result, long waitingTime, long executionTime, int zone) {
         this.serviceName = serviceName;
         this.result = result;
         this.waitingTime = waitingTime;
